# swagger.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/menuItems/{id}": {
      "get": {
        "tags": [
          "menu-controller"
        ],
        "summary": "Retrieve a single menu item by ID",
        "description": "Retrieves a menu item by its ID along with booking details.",
        "operationId": "getMenuItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "menu-controller"
        ],
        "summary": "Update an existing menu item",
        "description": "Updates an existing menu item by its ID, including its category.",
        "operationId": "updateMenuItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Menu item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "menu-controller"
        ],
        "summary": "Delete a menu item",
        "description": "Deletes a menu item by its ID.",
        "operationId": "deleteMenuItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Menu item deleted successfully"
          },
          "404": {
            "description": "Menu item not found"
          }
        }
      }
    },
...
    "/api/categories/{id}": {
      "get": {
        "tags": [
          "category-controller"
        ],
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "category-controller"
        ],
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "category-controller"
        ],
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get user by ID",
        "description": "Retrieves a user by their ID.",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Update user",
        "description": "Updates a user by their ID.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "Delete user",
        "description": "Deletes a user by their ID.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          }
        }
      }
    },
...
    "/admin/users/{userId}/role": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Update user role",
        "description": "Updates the role of a user based on the user ID.",
        "operationId": "updateUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User role updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      }
    },
    "/admin/user/{id}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get user by ID",
        "description": "Retrieves a user by their ID.",
        "operationId": "getUserById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Edit user details",
        "description": "Updates user details for the specified user ID.",
        "operationId": "editUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Delete a user",
        "description": "Deletes a user by their ID.",
        "operationId": "deleteUser_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      }
    },
    "/admin/menuItems/{id}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "updateMenuItem_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteMenuItem_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/categories/{id}": {
      "put": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "updateCategory_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "deleteCategory_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/menuItems": {
      "get": {
        "tags": [
          "menu-controller"
        ],
        "summary": "Retrieve all menu items",
        "description": "Retrieves all menu items from the database.",
        "operationId": "getAllMenuItems",
        "responses": {
          "200": {
            "description": "All menu items retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "menu-controller"
        ],
        "summary": "Add a new menu item",
        "description": "Creates a new menu item with a category.",
        "operationId": "addMenuItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": [
          "category-controller"
        ],
        "operationId": "getAllCategories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "category-controller"
        ],
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get all users",
        "description": "Retrieves all users.",
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user-controller"
        ],
        "summary": "Create a new user",
        "description": "Creates a new user.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/bookings/admin/user/{userId}": {
      "post": {
        "tags": [
          "booking-controller"
        ],
        "summary": "Create a booking for a specific user",
        "description": "Creates a booking for a user based on the user ID provided and sends an email notification.",
        "operationId": "createBookingForUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Booking created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error, failed to create booking",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/user": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Add a new user",
        "description": "Creates a new user with admin privileges.",
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, failed to create user",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      }
    },
    "/admin/menuItems": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getAllMenuItems_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "createMenuItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "summary": "Get all users",
        "description": "Retrieves all users.",
        "operationId": "getAllUsers_1",
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ROLE_ADMIN": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MenuItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "bookingIds": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "categoryName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookingDTO"
            }
          }
        }
      },
...
    }
  }
}